core: split visitors into several sub files
